+++ cod/tests/t6.c
@@ -58,7 +58,7 @@ main()
     static char extern_string[] = "int printf(string format, ...);";
     static cod_extern_entry externs[] =
     {
-        {"printf", (void*)(long)printf},
+        {"printf", (void*)(intptr_t)printf},
         {(void*)0, (void*)0}
     };
     static char code[] = "{\n\
@@ -83,7 +83,7 @@ main()
     cod_subroutine_declaration("int proc(FrameData *output)", context);
    
     gen_code = cod_code_gen(code, context);
-    func = (void (*)(void*))(long)gen_code->func;
+    func = (void (*)(void*))(intptr_t)gen_code->func;
 
     data.num_points = 0;
     data.image_data = NULL;
